Sorted out of the the dependencies issues.
Extension is loading, but fully funcional again yet.
