+++ injectors/FlexiInjector.py
@@ -1,4 +1,10 @@
 
+import os
+import urllib
+from helpers.FloydsHelpers import FloydsHelpers
+from injectors.Injector import Injector
+
+
 class FlexiInjector(Injector):
     # Can be used for any insertionPoint, as we simply globaly change stuff
     # with search/replace in the request.
